fix test

+++ tests/analyze.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
-rm outputs/tmp.csv
-python -u analyze.py tests/titanic/test_config.json > /dev/null
+if [ -e 'outputs/tmp.csv' ]; then
+  rm outputs/tmp.csv
+fi
+python -u analyze.py tests/titanic/test_config.json | grep ERROR
 diff_msg=$(diff outputs/tmp.csv tests/titanic/output.csv)
 if [ -n "${diff_msg}" ];then
   echo 'INTEGRATION TEST ERROR'
